Splitup index.js some, could still use more in routes, messages and login/signup working with req.session.passport
