+++ server/database/models/user.js
@@ -9,13 +9,13 @@ var UserSchema = new Schema({
     },
     password: String,
     profile: {
-    	name: String,
-    	about: String,
-    	interests: []
+      name: String,
+      about: String,
+      interests: []
     }
 });
+
 //hash
-//
 UserSchema.methods.generateHash = function(password) {
   return bcrypt.hashSync(password, bcrypt.genSaltSync(9));
 };
