+++ server/request_handler.js
@@ -25,15 +25,15 @@ exports.getUsers = function(req, res){
 
 //retrieves the profile information for the current user
 exports.getProfile = function(req, res){
-  //username is grabbed from request URL
-  var username = req.path.substring(7);
+  var userId = req.session.passport.user;
 
-  User.findOne({username: username}, function(err, user){
+  User.findOne({_id: userId}, function(err, user){
     if (err || !user){
       console.log('user does not exist');
       res.status(404).send(err);
       return;
     }
+    // console.log(user);
     res.status(200).send(user.profile);
   });
 };
@@ -52,9 +52,9 @@ exports.signUp = function(req, res) {
       console.log('failed to create new user');
       //eventually should be a redirect here
       res.status(400).send(err);
-      return; 
+      return;
     }
-    //eventually 
+    //eventually
     //should redirect to main
     //create a session
     res.status(201).send(user);
@@ -72,7 +72,7 @@ exports.signIn = function(req, res) {
       console.log('failed to find user');
       //probably should redirect or something here
       res.status(404).send(err);
-      return; 
+      return;
     }
     //should create session and perhaps redirect here
     res.status(200).send(user);
@@ -80,28 +80,29 @@ exports.signIn = function(req, res) {
 };
 
 //overwrites current profile information
-//will overwrite all three fields (name, about, interests) 
+//will overwrite all three fields (name, about, interests)
 exports.submitProfile = function(req, res){
   //grabs the username from request URL
-  var username = req.path.substring(7);
+  // var username = req.path.substring(7);
+  var userId = req.session.passport.user;
 
-  var profileInfo = {}; 
+  var profileInfo = {};
   profileInfo.name = req.body.name;
   profileInfo.about = req.body.about;
   profileInfo.interests = req.body.interests;
 
-  User.findOne({username: username}, function(err, user){
+  User.findOne({_id: userId}, function(err, user){
     if (err || !user){
       console.log('user does not exist');
       res.status(404).send(err);
-      return; 
+      return;
     }
     user.profile = profileInfo;
     user.save(function(err){
       if (err){
         console.log('profile was not updated');
         res.status(400).send(err);
-        return; 
+        return;
       }
 
       res.status(201).send(user.profile);
@@ -115,6 +116,9 @@ exports.sendMessage = function(req, res){
   messageInfo.from = req.body.from;
   messageInfo.text = req.body.text;
 
+  // console.log('messageInfo to', messageInfo.to);
+  // console.log('messageInfo from', messageInfo.from);
+  // console.log('messageInfo text', messageInfo.text);
   //TODO
   //verify that the sender exists
   //and has a session to prove who they are
@@ -126,6 +130,7 @@ exports.sendMessage = function(req, res){
       res.status(404).send(err);
       return;
     }
+    // console.log('recipient', recipient);
 
     //creates and saves the message
     var message = new Message(messageInfo);
@@ -135,49 +140,53 @@ exports.sendMessage = function(req, res){
         res.status(400).send(err);
         return;
       }
+      // console.log('message saved 143', message);
       res.status(201).send(message);
     });
   });
 };
 
-//finds all messages from a particular user and sorts them by conversation 
-//the shape of the data here is a bit convoluted and I apologize for that, 
+//finds all messages from a particular user and sorts them by conversation
+//the shape of the data here is a bit convoluted and I apologize for that,
 //but it makes life really easy on the client side in messenger.jsx
 //see README for a more precise description of what the server is sending
 exports.getMessages = function(req, res){
   //grabs username from URL
-  var username = req.path.substring(10);
+  // var username = req.path.substring(10);
+  var userId = req.session.passport.user;
   var allMessages = {};
 
- User.findOne({username: username}, function(err, user){
-  if (err || !user){
-    console.log('user does not exist');
-    res.status(404).send(err);
-    return;
-  }
+  User.findOne({_id: userId}, function(err, user){
+    if (err || !user){
+      console.log('user does not exist');
+      res.status(404).send(err);
+      return;
+    }
 
-  //TODO
-  //verify that this user is who they say they are
-  //via sessions or something
+    //TODO -- Done
+    //verify that this user is who they say they are
+    //via sessions or something
 
-    Message.find({'to': username}, function(err, receivedMessages){
+    Message.find({'to': user.username}, function(err, receivedMessages){
       if (err){
         console.log('could not find messages');
         res.status(404).send(err);
         return;
       }
 
-      receivedMessages.forEach(function(message){
-        var from = message.from;
+      if (Array.isArray(receivedMessages)) {
+        receivedMessages.forEach(function(message){
+          var from = message.from;
 
-        if (!allMessages[from]){
-          allMessages[from] = [];
-        }
+          if (!allMessages[from]){
+            allMessages[from] = [];
+          }
 
-        allMessages[from].push(message);
-      });
+          allMessages[from].push(message);
+        });
+      }
 
-      Message.find({'from': username}, function(err, sentMessages){
+      Message.find({'from': user.username}, function(err, sentMessages){
         if (err){
           console.log('could not find messages');
           res.status(404).send(err);
